refactor(web): move Layout logic to DesktopLayout refactor(web): add some i18n
