+++ ui/web/public/locales/zh/AgentConfForm.json
@@ -1,0 +1,17 @@
+{
+  "bundle": "打包",
+  "bundle_failed": "打包 {{ source }} 到 {{ target }} 失败",
+  "bundle_succeed": "打包 {{ source }} 到 {{ target }} 成功",
+  "load_config": "载入配置",
+  "load_config_failed": "加载 {{ filename }} 失败: {{ message }}",
+  "load_config_filename_prompt": "请输入配置文件路径",
+  "load_config_succeed": "加载 {{ filename }} 完成",
+  "refresh_schema": "刷新表单",
+  "require_config": "未找到模型配置，请先解析模型配置",
+  "require_entry_field": "需要补充模型入口路径字段",
+  "run": "运行",
+  "save_config": "保存配置",
+  "save_config_failed": "保存 {{ filename }} 失败",
+  "save_config_filename_prompt": "请输入保存配置的路径",
+  "save_config_succeed": "保存 {{ filename }} 成功"
+}
