+++ ui/web/src/modules/Agent/AgentConfForm.tsx
@@ -188,7 +188,7 @@ export const AgentConfForm = React.memo((props: { node?: TabNode }) => {
   const agentConf = useObservableState(agentConf$);
   const schema = useObservableState(agentConfSchema$) || {};
   const complete = useObservableState(complete$);
-  const { t } = useTranslation();
+  const { t } = useTranslation('AgentConfForm');
 
   useEffect(() => {
     if (props.node) {
@@ -202,14 +202,13 @@ export const AgentConfForm = React.memo((props: { node?: TabNode }) => {
         <Space style={{ width: '100%', flexWrap: 'wrap' }}>
           <Button
             icon={<IconPlay />}
-            title="运行代码"
             loading={!complete}
             onClick={() => {
               clearLogAction$.next();
               runAgentAction$.next();
             }}
           >
-            运行
+            {t('run')}
           </Button>
           <Button
             icon={<IconRefresh />}
@@ -217,65 +216,72 @@ export const AgentConfForm = React.memo((props: { node?: TabNode }) => {
               reloadSchemaAction$.next();
             }}
           >
-            刷新表单
+            {t('refresh_schema')}
           </Button>
           <Button
             icon={<IconFile />}
-            title="从文件载入配置"
             onClick={async () => {
+              const filename = prompt(t('load_config_filename_prompt'));
+              if (!filename) return;
               try {
-                const filename = prompt('配置文件路径');
-                if (filename) {
-                  const content = await fs.readFile(filename);
-                  const json = parse(content);
-                  agentConf$.next(json);
-                  Toast.success(`加载 ${filename} 完成`);
-                }
+                const content = await fs.readFile(filename);
+                const json = parse(content);
+                agentConf$.next(json);
+                Toast.success(t('load_config_succeed', { filename, interpolation: { escapeValue: false } }));
               } catch (e) {
-                Toast.error(`加载失败: ${e}`);
+                Toast.error(
+                  `${t('load_config_failed', { filename, interpolation: { escapeValue: false } })}: ${e}`,
+                );
               }
             }}
           >
-            载入配置
+            {t('load_config')}
           </Button>
           <Button
             icon={<IconSave />}
-            title="保存配置到文件"
             onClick={async () => {
+              if (!agentConf) return;
+              if (!agentConf.entry) return;
+              const filename = prompt(t('save_config_filename_prompt'));
+              if (!filename) return;
               try {
-                if (!agentConf) return;
-                if (!agentConf.entry) return;
-                const saveFilename = prompt('保存路径');
-                if (saveFilename) {
-                  const bundled_code = await bundleCode(agentConf.entry);
-                  await fs.writeFile(saveFilename, JSON.stringify({ ...agentConf, bundled_code }, null, 2));
-                  Toast.success(`保存到 ${saveFilename}`);
-                }
+                const bundled_code = await bundleCode(agentConf.entry);
+                await fs.writeFile(filename, JSON.stringify({ ...agentConf, bundled_code }, null, 2));
+                Toast.success(t('save_config_succeed', { filename, interpolation: { escapeValue: false } }));
               } catch (e) {
-                Toast.error(`保存失败: ${e}`);
+                Toast.error(
+                  `${t('save_config_failed', { filename, interpolation: { escapeValue: false } })}: ${e}`,
+                );
               }
             }}
           >
-            保存配置
+            {t('save_config')}
           </Button>
           <Button
             icon={<IconWrench />}
             onClick={async () => {
+              if (!agentConf) {
+                Toast.error(t('require_config'));
+                return;
+              }
+              if (!agentConf.entry) {
+                Toast.error(t('require_entry_field'));
+                return;
+              }
+              const source = agentConf.entry;
+              const target = `${source}.bundle.js`;
               try {
-                if (agentConf) {
-                  const agentCode = await bundleCode(agentConf.entry!);
-                  const bundleFilename = `${agentConf.entry}.bundle.js`;
-                  await fs.writeFile(bundleFilename, agentCode);
-                  Toast.success(`构建完成，保存到 ${bundleFilename}`);
-                } else {
-                  Toast.error(`请先解析模型配置`);
-                }
+                const agentCode = await bundleCode(source);
+                await fs.writeFile(target, agentCode);
+                Toast.success(t('bundle_succeed', { source, target, interpolation: { escapeValue: false } }));
               } catch (e) {
-                Toast.error(`保存失败: ${e}`);
+                Toast.error(
+                  `${t('bundle_failed', { source, target, interpolation: { escapeValue: false } })}: ${e}`,
+                );
               }
             }}
           >
-            构建
+            {t('bundle')}
           </Button>
         </Space>
         <Divider />
